Wait for app ready to delete old Semeru instances (#349)
* Wait for app ready to delete old Semeru instances

* Update go.mod to pull runtime-component-operator#438

* Update isWebSphereLibertyApplicationReady

* Check for areCompletedSemeruInstancesMarkedToBeDeleted before app status

* Remove extra RCO versions in go.sum

+++ controllers/webspherelibertyapplication_controller.go
@@ -658,7 +658,7 @@ func (r *ReconcileWebSphereLiberty) Reconcile(ctx context.Context, request ctrl.
 	}
 
 	// Delete completed Semeru instances because all pods now point to the newest Semeru service
-	if areCompletedSemeruInstancesMarkedToBeDeleted {
+	if areCompletedSemeruInstancesMarkedToBeDeleted && r.isWebSphereLibertyApplicationReady(instance) {
 		if err := r.deleteCompletedSemeruInstances(instance); err != nil {
 			reqLogger.Error(err, "Failed to delete completed Semeru instance")
 			return r.ManageError(err, common.StatusConditionTypeReconciled, instance)
@@ -668,6 +668,14 @@ func (r *ReconcileWebSphereLiberty) Reconcile(ctx context.Context, request ctrl.
 	instance.Status.Versions.Reconciled = lutils.OperandVersion
 	reqLogger.Info("Reconcile WebSphereLibertyApplication - completed")
 	return r.ManageSuccess(common.StatusConditionTypeReconciled, instance)
+}
+
+func (r *ReconcileWebSphereLiberty) isWebSphereLibertyApplicationReady(ba common.BaseComponent) bool {
+	if r.CheckApplicationStatus(ba) == corev1.ConditionTrue {
+		statusCondition := ba.GetStatus().GetCondition(common.StatusConditionTypeReady)
+		return statusCondition != nil && statusCondition.GetMessage() == common.StatusConditionTypeReadyMessage
+	}
+	return false
 }
 
 func (r *ReconcileWebSphereLiberty) SetupWithManager(mgr ctrl.Manager) error {
